Separate regression logic
